+++ app/controllers/concerns/folio/set_current_request_details.rb
@@ -13,7 +13,8 @@ module Folio::SetCurrentRequestDetails
         # warden takes params from request, so we need to set source_site_id here,
         # before searching for current user
         if request.params["user"].present? && request.params["user"]["auth_site_id"].blank?
-          request.params["user"]["auth_site_id"] = Folio::Current.site.id.to_s
+          # don't use `.site` here as that would try to fetch one without host and cache it
+          request.params["user"]["auth_site_id"] = Folio::Current.get_site(host: request.host).id.to_s
         end
 
         Folio::Current.setup!(request:,