+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 - `react_ordered_multiselect` can now show button to fix required field
+- setting of `auth_site_id` caching wrong site in `Folio::Current`
 
 ## [6.3.0] - 2025-04-03